+++ s/rand/index.ts
@@ -1,0 +1,6 @@
+
+export * from "./rand.js"
+export * from "./seed.js"
+export * from "./types.js"
+export * from "./u32ify.js"
+