+++ s/rand/types.ts
@@ -1,0 +1,4 @@
+
+/** fn that returns random number fraction between 0 and 1. */
+export type Random = () => number
+